Move frontier, program, type, utilities modules under lib/ and update imports
