+++ demo1.py
@@ -3,10 +3,10 @@ import os
 
 from lib.ec import commandlineArguments, ecIterator
 from grammar import Grammar
-from program import Primitive
+from lib.program import Primitive
 from lib.task import Task
-from type import t0, arrow, tlist, tint, tbool
-from utilities import numberOfCPUs
+from lib.type import t0, arrow, tlist, tint, tbool
+from lib.utilities import numberOfCPUs
 
 # input/output types
 intlist = arrow(tlist(tint), tlist(tint))
