+++ lib/__init__.py
@@ -24,9 +24,13 @@ For more info, see this StackOverflow answer: https://stackoverflow.com/a/212191
 import sys
 
 from lib import ec
+from lib import frontier
+from lib import program
 from lib import primitiveGraph
 from lib import task
 from lib import taskBatcher
+from lib import type
+from lib import utilities
 from lib.domains.misc import algolispPrimitives, deepcoderPrimitives
 from lib.domains.misc import RobustFillPrimitives
 from lib.domains.misc import napsPrimitives
@@ -43,8 +47,12 @@ from lib.domains.text import textPrimitives
 from lib.domains.text import makeTextTasks
 
 sys.modules['ec'] = ec
+sys.modules['frontier'] = frontier
+sys.modules['program'] = program
 sys.modules['task'] = task
 sys.modules['taskBatcher'] = taskBatcher
+sys.modules['type'] = type
+sys.modules['utilities'] = utilities
 sys.modules['algolispPrimitives'] = algolispPrimitives
 sys.modules['RobustFillPrimitives'] = RobustFillPrimitives
 sys.modules['napsPrimitives'] = napsPrimitives
