+++ lib/domains/misc/algolispPrimitives.py
@@ -1,9 +1,8 @@
 #napsPrimitives.py
-from program import Primitive, Program, prettyProgram
+from lib.program import Primitive, Program
 from grammar import Grammar
-from type import tlist, tint, tbool, arrow, baseType #, t0, t1, t2
+from lib.type import tlist, arrow, baseType #, t0, t1, t2
 
-import math
 #from functools import reduce
