+++ lib/domains/puddleworld/puddleworldPrimitives.py
@@ -6,12 +6,6 @@ Converts a PyCCG ontology into primitives for Dreamcoder.
 import sys
 sys.path.insert(0, "pyccg")
 sys.path.insert(0, "pyccg/nltk")
-from pyccg.logic import TypeSystem, Ontology, Expression
-from puddleworld.ontology import puddleworldOntology
-
-from type import baseType
-from utilities import eprint
-
 
 """
 tLayoutMap = baseType('layoutMap')
