+++ lib/type.py
