+++ lib/utilities.py
