+++ list.py
@@ -6,10 +6,10 @@ import os
 import datetime
 
 from lib.ec import explorationCompression, commandlineArguments
-from utilities import eprint, numberOfCPUs, flatten, testTrainSplit
+from lib.utilities import eprint, numberOfCPUs, flatten, testTrainSplit
 from grammar import Grammar
 from lib.task import Task
-from type import Context, arrow, tbool, tlist, tint, t0, UnificationFailure
+from lib.type import Context, arrow, tbool, tlist, tint, t0, UnificationFailure
 from lib.domains.list.listPrimitives import basePrimitives, primitives, McCarthyPrimitives, bootstrapTarget_extra, no_length
 from recognition import RecurrentFeatureExtractor
 from lib.domains.list.makeListTasks import make_list_bootstrap_tasks, sortBootstrap
