+++ python_server.py
@@ -1,4 +1,4 @@
-from utilities import *
+from lib.utilities import *
 
 import threading
 import zmq
