+++ regexes.py
@@ -2,17 +2,17 @@
 
 from lib.ec import explorationCompression, commandlineArguments, Task
 from grammar import Grammar
-#from utilities import eprint, testTrainSplit, numberOfCPUs, flatten
-from utilities import eprint, numberOfCPUs, flatten, testTrainSplit, POSITIVEINFINITY
+#from lib.utilities import eprint, testTrainSplit, numberOfCPUs, flatten
+from lib.utilities import eprint, numberOfCPUs, flatten, testTrainSplit, POSITIVEINFINITY
 from lib.domains.regex.makeRegexTasks import makeOldTasks, makeLongTasks, makeShortTasks, makeWordTasks, makeNumberTasks, makeHandPickedTasks, makeNewTasks, makeNewNumberTasks
 from lib.domains.regex.regexPrimitives import basePrimitives, altPrimitives, easyWordsPrimitives, alt2Primitives, concatPrimitives, reducedConcatPrimitives, strConstConcatPrimitives
 from likelihoodModel import add_cutoff_values, add_string_constants
 
-from program import Abstraction, Application
+from lib.program import Abstraction, Application
 #from program import *
 from recognition import RecurrentFeatureExtractor, JSONFeatureExtractor
 import random
-from type import tpregex
+from lib.type import tpregex
 import math
 import pregex as pre
 import os
