+++ tower.py
@@ -3,7 +3,7 @@ from lib.ec import *
 from lib.domains.tower.towerPrimitives import primitives, new_primitives, animateTower
 from lib.domains.tower.makeTowerTasks import *
 from tower_common import renderPlan, towerLength, centerTower
-from utilities import *
+from lib.utilities import *
 
 import os
 import datetime
